+++ pytest_inmanta/plugin.py
@@ -501,10 +501,13 @@ class MockAgent(object):
 
     def __init__(self, uri):
         self.uri = uri
-        self.process = MockProcess()
         self._env_id = cfg_env.get()
         self.sessionid = "mockid"
         self.environment = self._env_id
+        # This is for the new old api in inmanta.agent.agent.AgentInstance
+        self.process = MockProcess()
+        # This is for the new agent api in inmanta.agent.executor.AgentInstance
+        self.eventloop = self.process._io_loop
 
 
 class MockClient(object):
@@ -543,9 +546,9 @@ class InmantaPluginsImportLoader:
         self._importer: InmantaPluginsImporter = importer
 
     def __getattr__(self, name: str):
-        submodules: typing.Optional[typing.Dict[str, ModuleType]] = (
-            self._importer.get_submodules(name)
-        )
+        submodules: typing.Optional[
+            typing.Dict[str, ModuleType]
+        ] = self._importer.get_submodules(name)
         fq_mod_name: str = f"inmanta_plugins.{name}"
         if submodules is None or fq_mod_name not in submodules:
             raise AttributeError("No inmanta module named %s" % name)
@@ -871,9 +874,9 @@ class Project:
         self._root_scope: typing.Optional[inmanta.ast.Namespace] = None
         self._exporter: typing.Optional[Exporter] = None
         self._blobs: typing.Dict[str, bytes] = {}
-        self._facts: typing.Dict[ResourceIdStr, typing.Dict[str, typing.Any]] = (
-            defaultdict(dict)
-        )
+        self._facts: typing.Dict[
+            ResourceIdStr, typing.Dict[str, typing.Any]
+        ] = defaultdict(dict)
         self._should_load_plugins: typing.Optional[bool] = load_plugins
         self._plugins: typing.Optional[typing.Dict[str, FunctionType]] = None
         self._load()
@@ -1388,9 +1391,9 @@ license: Test License
     def _load_plugins(self) -> typing.Dict[str, FunctionType]:
         mod: module.Module
         mod, _ = get_module()
-        submodules: typing.Optional[typing.Dict[str, ModuleType]] = (
-            InmantaPluginsImporter(self).get_submodules(mod.name)
-        )
+        submodules: typing.Optional[
+            typing.Dict[str, ModuleType]
+        ] = InmantaPluginsImporter(self).get_submodules(mod.name)
         return (
             {}
             if submodules is None